refactor(bridge): :zap: move interpretter to specific folder
