+++ packages/bridge/src/eos/EOSTypeInterpreter.ts
@@ -1,17 +1,17 @@
+import { AbsTypeInterpreter } from '../interpreter/AbsTypeInterpreter';
 import { FieldTypeEnum } from '../type-definition/FieldTypeEnum';
-import { AbsTypeInterpreter } from './AbsTypeInterpreter';
 
-export class ICONTypeInterpreter extends AbsTypeInterpreter {
+export class EOSTypeInterpreter extends AbsTypeInterpreter {
   interpret(type: FieldTypeEnum): string {
     let mappedType: string;
 
     switch (type) {
       case FieldTypeEnum.number:
       case FieldTypeEnum.uint64_t:
-        mappedType = 'int';
+        mappedType = 'uint64_t';
         break;
       case FieldTypeEnum.string:
-        mappedType = 'str';
+        mappedType = 'std::string';
         break;
       default:
         throw new Error(`unknow type ${type}`);
