+++ packages/bridge/src/icon/ICONContractAdapter.ts
@@ -3,11 +3,11 @@ import Handlebars from 'handlebars';
 import path from 'path';
 
 import { AbsContractAdapter } from '../AbsContractAdapter';
-import { ICONTypeInterpreter } from '../interpreter/ICONTypeInterpreter';
-import { FieldTypeEnum } from '../type-definition/FieldTypeEnum';
 import { JsPrettier } from '../prettier/JsPrettier';
 import { PythonPrettier } from '../prettier/PythonPrettier';
 import { FilePrinter } from '../printer/FilePrinter';
+import { FieldTypeEnum } from '../type-definition/FieldTypeEnum';
+import { ICONTypeInterpreter } from './ICONTypeInterpreter';
 
 import type { Printer } from '../printer/Printer';
 
@@ -42,7 +42,10 @@ export class ICONContractAdapter extends AbsContractAdapter {
 
   generateInit() {
     const fileName = '__init__.py';
-    const hbsTemplate = fs.readFileSync(path.resolve(this.templatePath, this.folderName, `${fileName}.hbs`), 'utf-8');
+    const hbsTemplate = fs.readFileSync(
+      path.resolve(this.templatePath, this.folderName, `${fileName}.hbs`),
+      'utf-8',
+    );
     const template = Handlebars.compile(hbsTemplate);
     const outText = template({});
     this.jsFilePrinter.print(fileName, outText, this.folderName);
@@ -50,7 +53,10 @@ export class ICONContractAdapter extends AbsContractAdapter {
 
   generatePackage() {
     const fileName = 'package.json';
-    const hbsTemplate = fs.readFileSync(path.resolve(this.templatePath, this.folderName, `${fileName}.hbs`), 'utf-8');
+    const hbsTemplate = fs.readFileSync(
+      path.resolve(this.templatePath, this.folderName, `${fileName}.hbs`),
+      'utf-8',
+    );
     const template = Handlebars.compile(hbsTemplate);
 
     // translate
@@ -59,14 +65,16 @@ export class ICONContractAdapter extends AbsContractAdapter {
     this.jsFilePrinter.print(fileName, outText, this.folderName);
   }
 
-
   generateMainPy() {
     const fileName = 'icon_hello.py';
-    const hbsTemplate = fs.readFileSync(path.resolve(this.templatePath, this.folderName, `${fileName}.hbs`), 'utf-8');
+    const hbsTemplate = fs.readFileSync(
+      path.resolve(this.templatePath, this.folderName, `${fileName}.hbs`),
+      'utf-8',
+    );
     const template = Handlebars.compile(hbsTemplate);
 
     const outText = template({
-      tables: this.entityConfigs ?.map(item => {
+      tables: this.entityConfigs?.map(item => {
         return {
           name: item.name,
           fields: item.fields.map(({ name, type }) => ({
@@ -81,14 +89,17 @@ export class ICONContractAdapter extends AbsContractAdapter {
 
   generateCallAPI() {
     const fileName = 'call.json';
-    const hbsTemplate = fs.readFileSync(path.resolve(this.templatePath, this.folderName, `${fileName}.hbs`), 'utf-8');
+    const hbsTemplate = fs.readFileSync(
+      path.resolve(this.templatePath, this.folderName, `${fileName}.hbs`),
+      'utf-8',
+    );
     const template = Handlebars.compile(hbsTemplate);
 
-    this.entityConfigs ?.map(item => {
-      const actionName = "get" + item.name
-      const outFileName = actionName + ".json";
+    this.entityConfigs?.map(item => {
+      const actionName = 'get' + item.name;
+      const outFileName = actionName + '.json';
       const outText = template({
-        action: actionName.toLocaleLowerCase()
+        action: actionName.toLocaleLowerCase(),
       });
       this.jsFilePrinter.print(outFileName, outText, this.folderName);
     });
@@ -96,13 +107,16 @@ export class ICONContractAdapter extends AbsContractAdapter {
 
   generateTXAPI() {
     const fileName = 'tx.json';
-    const hbsTemplate = fs.readFileSync(path.resolve(this.templatePath, this.folderName, `${fileName}.hbs`), 'utf-8');
+    const hbsTemplate = fs.readFileSync(
+      path.resolve(this.templatePath, this.folderName, `${fileName}.hbs`),
+      'utf-8',
+    );
     const template = Handlebars.compile(hbsTemplate);
 
-    this.entityConfigs ?.map(item => {
+    this.entityConfigs?.map(item => {
       // create data
-      const crtActionName = "crt" + item.name
-      const crtFileName = crtActionName + ".json";
+      const crtActionName = 'crt' + item.name;
+      const crtFileName = crtActionName + '.json';
       const crtText = template({
         action: crtActionName.toLocaleLowerCase(),
         fields: item.fields.map(({ name, type }) => ({
@@ -113,8 +127,8 @@ export class ICONContractAdapter extends AbsContractAdapter {
       this.jsFilePrinter.print(crtFileName, crtText, this.folderName);
 
       // update data
-      const udtActionName = "upt" + item.name
-      const uptFileName = udtActionName + ".json";
+      const udtActionName = 'upt' + item.name;
+      const uptFileName = udtActionName + '.json';
       const uptText = template({
         action: udtActionName.toLocaleLowerCase(),
         fields: item.fields.map(({ name, type }) => ({
@@ -125,14 +139,13 @@ export class ICONContractAdapter extends AbsContractAdapter {
       this.jsFilePrinter.print(uptFileName, uptText, this.folderName);
 
       // remove data
-      const rmvActionName = "rmv" + item.name
-      const rmvFileName = rmvActionName + ".json";
+      const rmvActionName = 'rmv' + item.name;
+      const rmvFileName = rmvActionName + '.json';
       const rmvText = template({
         action: rmvActionName.toLocaleLowerCase(),
-        fields: [{"name": "id", "type": "int"}]
+        fields: [{ name: 'id', type: 'int' }],
       });
       this.jsFilePrinter.print(rmvFileName, rmvText, this.folderName);
-
     });
   }
   generateUnitTestPy() {
