+++ packages/bridge/src/icon/ICONTypeInterpreter.ts
@@ -1,17 +1,17 @@
+import { AbsTypeInterpreter } from '../interpreter/AbsTypeInterpreter';
 import { FieldTypeEnum } from '../type-definition/FieldTypeEnum';
-import { AbsTypeInterpreter } from './AbsTypeInterpreter';
 
-export class EOSTypeInterpreter extends AbsTypeInterpreter {
+export class ICONTypeInterpreter extends AbsTypeInterpreter {
   interpret(type: FieldTypeEnum): string {
     let mappedType: string;
 
     switch (type) {
       case FieldTypeEnum.number:
       case FieldTypeEnum.uint64_t:
-        mappedType = 'uint64_t';
+        mappedType = 'int';
         break;
       case FieldTypeEnum.string:
-        mappedType = 'std::string';
+        mappedType = 'str';
         break;
       default:
         throw new Error(`unknow type ${type}`);
